chore: move all pipe logic to pipe module
