+++ pipe.d.ts
