+++ src/index.ts
@@ -9,5 +9,5 @@ export * as collections from './collections/index.js';
 export * as events from './events/index.js';
 export * as operators from './operators/index.js';
 export * as refs from './refs/index.js';
-export * as steps from './steps/index.js';
+export * as steps from './pipe/index.js';
 export * as subscriptions from './subscriptions/index.js';
