+++ src/operators/index.ts
@@ -1,7 +1,5 @@
 export * from './dynamic.js';
-export * from './flow.js';
 export * from './inherit.js';
 export * from './lazy.js';
 export * from './merge.js';
 export * from './pick.js';
-export * from './pipe.js';
