+++ src/pipe/index.ts
@@ -1,5 +1,7 @@
 export * from './cache.js';
 export * from './dedupe-read.js';
 export * from './each.js';
+export * from './flow.js';
 export * from './filter.js';
 export * from './json.js';
+export * from './pipe.js';
