+++ src/pipe/pipe.ts
