+++ tests/pipe/flow.test.ts
@@ -1,8 +1,8 @@
 import { vi } from 'vitest';
 
 import { Observable } from '@/src/defs/features/observable.js';
-import { flow$ } from '@/src/operators/flow.js';
-import { PipeContext } from '@/src/operators/pipe.js';
+import { flow$ } from '@/src/pipe/flow.js';
+import { PipeContext } from '@/src/pipe/pipe.js';
 import { source$ } from '@/src/source.js';
 
 describe('flow$', () => {
