+++ tests/pipe/json.test-d.ts
@@ -1,11 +1,11 @@
 /* eslint-disable @typescript-eslint/no-unused-vars */
 import { describe, expectTypeOf } from 'vitest';
 
-import { pipe$ } from '@/src/operators/pipe.js';
+import { pipe$ } from '@/src/pipe/pipe.js';
 import { const$ } from '@/src/refs/const.js';
 import { ref$ } from '@/src/refs/ref.js';
 import { var$ } from '@/src/refs/var.js';
-import { json$ } from '@/src/steps/index.js';
+import { json$ } from '@/src/pipe/index.js';
 import { source$ } from '@/src/source.js';
 
 // Tests
